setup_requirements.txt
kafka-python
faker
python-dotenv==1.0.1
confluent-kafka==2.3.0
numpy<2
pyarrow==14.0.2